+++ src/main/java/com/exclamationlabs/connid/base/ringcentral/RingCentralConnector.java
@@ -17,9 +17,14 @@ import com.exclamationlabs.connid.base.connector.BaseFullAccessConnector;
 import com.exclamationlabs.connid.base.connector.authenticator.OAuth2TokenPasswordAuthenticator;
 import com.exclamationlabs.connid.base.ringcentral.adapter.RingCentralCallQueueAdapter;
 import com.exclamationlabs.connid.base.ringcentral.adapter.RingCentralUsersAdapter;
+import com.exclamationlabs.connid.base.ringcentral.attribute.RingCentralCallQueueAttribute;
+import com.exclamationlabs.connid.base.ringcentral.attribute.RingCentralUserAttribute;
 import com.exclamationlabs.connid.base.ringcentral.configuration.RingCentralConfiguration;
 import com.exclamationlabs.connid.base.ringcentral.driver.rest.RingCentralDriver;
 import org.identityconnectors.framework.spi.ConnectorClass;
+
+import java.util.HashSet;
+import java.util.Set;
 
 @ConnectorClass(displayNameKey = "ringcentral.connector.display", configurationClass = RingCentralConfiguration.class)
 
@@ -30,6 +35,11 @@ public class RingCentralConnector extends BaseFullAccessConnector {
 
         setDriver(new RingCentralDriver());
         setAdapters(new RingCentralUsersAdapter(), new RingCentralCallQueueAdapter());
+        setEnhancedFiltering(true);
+        Set<String> filterSet = new HashSet<>();
+        filterSet.add(RingCentralUserAttribute.USER_NAME.name());
+        filterSet.add(RingCentralCallQueueAttribute.USER_MEMBERS.name());
+        setFilterAttributes(filterSet);
     }
 
 }