+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Change Log
 
++ **0.5** - Added filtering support for call queue for user id EAGLE-330 (2021/11/09)
++ **0.4** - Added filtering support for userName EAGLE-330 (2021/11/09)
 + **0.3** - Added support for call queues EAGLE-330 (2021/11/09)
 + **0.2** - Upgrade to base connector 1.1.5 (2021/09/07)
 + **0.1** - Initial development (2021/06/17)